+++ transpiler/SinusbotScriptFunctions.js
@@ -1,0 +1,8 @@
+/**
+ * @param {string} path
+ */
+export function importClass(path) {}
+/**
+ * @param {string} path
+ */
+export function importObject(path) {}
